requirements-dev.txt
pre-commit==0.14.2
isort==4.2.5
django-extensions
psycopg2
git+ssh://git@github.com/shimizukawa/django-redshift-backend.git@master
ipython
ipdb
flower
honcho
